+++ src/main/java/com/example/jsgamesbackendmain/Service/RankService.java
@@ -1,0 +1,2 @@
+package com.example.jsgamesbackendmain.Service;public class RankService {
+}
